Fix that self.duration nonsense and report on error
Part of the problem when it falls over is that the users are not technical
enough to give accurate error reports. Hopefully by capturing what the
actual kaboom is, it will enable more defensive programming.

If you can call assigning int values to a string-initialized member variable
as "programming". Just damn.